+++ src/pyscaffold/extensions/pre_commit.py
@@ -19,7 +19,7 @@ def augment_cli(parser):
     """
 
     parser.add_argument(
-        "--with-pre-commit",
+        "--pre-commit",
         dest="extensions",
         action="append_const",
         const=extend_project,