+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Version 3.0, 2017-??-??
 - Updated setuptools_scm to master (SET TO ACTUAL RELEASE TAG!)
 - Changed my_project/my_package to recommended my_project/src/my_package structure
 - Renamed CHANGES.rst to more standard CHANGELOG.rst
+- Added platforms parameter in setup.cfg
 
 Version 2.5.8, 2017-09-10
 =========================